Moved Models code
