+++ main.py
@@ -13,7 +13,8 @@ from contextlib import asynccontextmanager
 from fastapi import FastAPI, HTTPException, Request, status, Depends
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.responses import JSONResponse
-from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
+from fastapi.security import     length: float = Field(0, ge=0, description="Package length")
+HTTPBearer, HTTPAuthorizationCredentials
 from pydantic import BaseModel, Field, validator
 import uvicorn
 from routers import all_routers
@@ -37,198 +38,6 @@ try:
 except Exception as e:
     logging.getLogger().setLevel(logging.INFO)
     logger.error(f"Error setting log level: {e}, defaulting to INFO")
-
-
-# Request/Response Models
-class Address(BaseModel):
-    # Required fields
-    postal_code: str = Field(..., description="Postal/ZIP code")
-    country_code: str = Field(..., description="Country code")
-    
-    # Optional fields
-    city: Optional[str] = Field(None, description="City name")
-    state_province_code: Optional[str] = Field(None, description="State/Province code")
-    address_line: Optional[str] = Field(None, description="Street address")
-    
-    @validator('country_code')
-    def validate_country_code(cls, v):
-        return v.upper()
-
-class Package(BaseModel):
-    # Required
-    weight: float = Field(..., gt=0, description="Package weight")
-    
-    # Optional dimensions
-    length: float = Field(0, ge=0, description="Length")
-    width: float = Field(0, ge=0, description="Width") 
-    height: float = Field(0, ge=0, description="Height")
-    
-    # Optional fields with Ruby defaults
-    packaging_type: str = Field("your_packaging", description="Package type")
-    insured_value: float = Field(0, ge=0, description="Insured value")
-    currency_code: str = Field("USD", description="Currency code")
-    
-    @validator('packaging_type')
-    def validate_packaging_type(cls, v):
-        if v not in UPSConstants.PACKAGES:
-            logger.warning(f"Unknown packaging type: {v}, using default 'your_packaging'")
-            return "your_packaging"
-        return v
-    
-    @validator('weight')
-    def validate_weight_minimum(cls, v):
-        # UPS requirement: minimum 1.0 pound
-        return max(1.0, float(v))
-
-class UPSRateRequest(BaseModel):
-    """Complete UPS Rate Request matching Ruby UPSRateRequest class"""
-    
-    # Required fields
-    sender_country: str = Field(..., description="Sender country code")
-    country: str = Field(..., description="Destination country code")  
-    weight: float = Field(..., gt=0, description="Package weight")
-    
-    # Conditional required fields (zip codes required for domestic US)
-    zip: Optional[str] = Field(None, description="Destination ZIP code")
-    sender_zip: Optional[str] = Field(None, description="Sender ZIP code")
-    
-    # Optional fields with Ruby defaults
-    service: str = Field("all", description="Service type")
-    package: str = Field("your_packaging", description="Package type")
-    length: float = Field(0, ge=0, description="Package length")
-    width: float = Field(0, ge=0, description="Package width")
-    height: float = Field(0, ge=0, description="Package height")
-    customer_type: str = Field("wholesale", description="Customer type")
-    pickup_type: str = Field("daily_pickup", description="Pickup type")
-    weight_units: str = Field("LBS", description="Weight units")
-    measure_units: str = Field("IN", description="Dimension units")
-    
-    # Optional address fields
-    city: Optional[str] = Field(None, description="Destination city")
-    state: Optional[str] = Field(None, description="Destination state")
-    sender_city: Optional[str] = Field(None, description="Sender city")
-    sender_state: Optional[str] = Field(None, description="Sender state")
-    
-    # Optional service options
-    insured_value: float = Field(0, ge=0, description="Insured value")
-    currency_code: str = Field("USD", description="Currency code")
-    
-
-    
-    @validator('service')
-    def validate_service(cls, v):
-        if v not in UPSConstants.SERVICES:
-            logger.warning(f"Unknown service: {v}, using default 'all'")
-            return "all"
-        return v
-    
-    @validator('package')
-    def validate_package(cls, v):
-        if v not in UPSConstants.PACKAGES:
-            logger.warning(f"Unknown package type: {v}, using default 'your_packaging'")
-            return "your_packaging"
-        return v
-    
-    @validator('customer_type')
-    def validate_customer_type(cls, v):
-        if v not in UPSConstants.CUSTOMER_TYPES:
-            logger.warning(f"Unknown customer type: {v}, using default 'wholesale'")
-            return "wholesale"
-        return v
-    
-    @validator('pickup_type')  
-    def validate_pickup_type(cls, v):
-        if v not in UPSConstants.PICKUP_TYPES:
-            logger.warning(f"Unknown pickup type: {v}, using default 'daily_pickup'")
-            return "daily_pickup"
-        return v
-    
-    def is_international(self) -> bool:
-        """Check if shipment is international"""
-        return self.sender_country != "US" or self.country != "US"
-    
-    def is_domestic_us(self) -> bool:
-        """Check if shipment is domestic US"""
-        return self.sender_country == "US" and self.country == "US"
-    
-    def validate_required_fields(self):
-        """Enhanced validation for all shipping scenarios based on UPS SDK patterns"""
-        errors = []
-        warnings = []
-        
-        # Basic required fields for all shipments
-        required_basic = ["sender_country", "country", "weight"]
-        for field in required_basic:
-            if not getattr(self, field, None):
-                errors.append(f"Missing required field: {field}")
-        
-        # Country code validation
-        if self.sender_country and len(self.sender_country) != 2:
-            errors.append("sender_country must be a 2-character ISO country code")
-        if self.country and len(self.country) != 2:
-            errors.append("country must be a 2-character ISO country code")
-        
-        # Domestic US shipments - flexible requirements
-        if self.is_domestic_us():
-            # ZIP codes are mandatory for US domestic
-            if not self.zip or not self.sender_zip:
-                errors.append("ZIP codes (zip and sender_zip) are required for US domestic shipments")
-            
-            # Cities are recommended for US domestic (not mandatory to avoid blocking valid requests)
-            if not self.city:
-                warnings.append("Destination city is recommended for US domestic shipments for better accuracy")
-            if not self.sender_city:
-                warnings.append("Sender city is recommended for US domestic shipments for better accuracy")
-            
-            # States are recommended for US domestic
-            if not self.state:
-                warnings.append("Destination state is recommended for US domestic shipments")
-            if not self.sender_state:
-                warnings.append("Sender state is recommended for US domestic shipments")
-        
-        # International shipments - flexible requirements
-        elif self.is_international():
-            # Postal codes are strongly recommended for international
-            if not self.zip:
-                warnings.append("Destination postal code is strongly recommended for international shipments")
-            if not self.sender_zip:
-                warnings.append("Sender postal code is strongly recommended for international shipments")
-            
-            # Cities are recommended for international
-            if not self.city:
-                warnings.append("Destination city is recommended for international shipments")
-            if not self.sender_city:
-                warnings.append("Sender city is recommended for international shipments")
-        
-        # Package dimensions validation
-        if any([self.length, self.width, self.height]) and not all([self.length > 0, self.width > 0, self.height > 0]):
-            warnings.append("If providing dimensions, all three (length, width, height) should be greater than 0")
-        
-        # Log warnings
-        for warning in warnings:
-            logger.warning(warning)
-        
-        # Raise errors if any
-        if errors:
-            error_msg = "; ".join(errors)
-            raise ValueError(error_msg)
-
-class RateQuote(BaseModel):
-    service_code: str
-    service_name: str
-    total_charge: float
-    currency_code: str
-    delivery_days: Optional[int] = None
-    guaranteed: Optional[bool] = None
-    context: Optional[str] = None
-    method_id: Optional[int] = None
-
-class RateResponse(BaseModel):
-    quotes: List[RateQuote]
-    request_id: str
-    timestamp: datetime
-    processing_time_ms: float
-    context: str
 
 # Enhanced UPS Rating Service
 class UPSRatingService:
